Set CORS to allow any origin.
If possible, I should restrict this later on, but I will probably not. It is probably worth to set it on a production server, but as this is just a simple course. It would be really bad if I set this now and then I mess up my presentation because of it being hosted on a different location.