Implement jwt tokens.
Makes for a more seamless experience when you come back to the page.
Previously you had to manually login again.

I probably spent around 20 hours on getting this to work, which in the end turned out not to work because of a typo, so the code is likely a mess.